fix: 빌드 오류 및 충돌 수정 (KGU-25)
