+++ api/src/main/java/uk/jinhy/server/api/vaccination/presentation/VaccinationDto.java
@@ -4,12 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import uk.jinhy.server.service.domain.VaccinationEntity;
-
 
 import java.time.LocalDate;
 import java.util.List;
-
 
 public class VaccinationDto {
 
@@ -43,17 +40,5 @@ public class VaccinationDto {
     public static class VaccinationListResponse {
         private List<VaccinationResponse> vaccinations;
         private int total;
-    }
-
-    // Entity → DTO 변환 메서드
-    public static VaccinationResponse fromEntity(VaccinationEntity entity) {
-        return VaccinationResponse.builder()
-            .id(entity.getId())  // 백신 기록 ID
-            .petId(entity.getPet().getId())  // 반려동물 ID
-            .vaccineName(entity.getVaccineName())  // 백신 이름
-            .vaccinationDate(entity.getVaccinationDate())  // 접종 날짜
-            .nextVaccinationDate(entity.getNextVaccinationDate())  // 다음 접종 날짜
-            .isCompleted(entity.isCompleted())  // 접종 완료 여부
-            .build();
     }
 }
