--- service/src/main/java/uk/jinhy/server/service/user/domain/UserEntityRepository.java
@@ -1,11 +1,0 @@
-package uk.jinhy.server.service.user.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
-
-    Optional<UserEntity> findByOauth2UserId(String oAuth2User);
-
-}
