+++ service/src/main/java/uk/jinhy/server/service/user/domain/UserRepository.java
@@ -8,4 +8,5 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByUsername(String username);
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
+    Optional<UserEntity> findByOauth2UserId(String oAuth2User);
 }
