+++ service/src/main/java/uk/jinhy/server/service/user/oauth2/CustomOAuth2UserService.java
@@ -1,6 +1,5 @@
 package uk.jinhy.server.service.user.oauth2;
 
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
@@ -8,14 +7,14 @@ import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import uk.jinhy.server.service.user.domain.UserEntityRepository;
+import uk.jinhy.server.service.user.domain.UserRepository;
 
 @Slf4j
 @RequiredArgsConstructor
 @Service
 public class CustomOAuth2UserService extends DefaultOAuth2UserService {
 
-    private final UserEntityRepository userEntityRepository;
+    private final UserRepository userRepository;
 
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
@@ -25,8 +24,8 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
         OAuth2UserResponse oAuth2UserResponse = OAuth2UserResponseFactory.create(oAuth2User, registrationId);
 
         String oAuth2UserId = oAuth2UserResponse.getOAuth2UserId();
-        userEntityRepository.findByOauth2UserId(oAuth2UserId)
-            .orElseGet(() -> userEntityRepository.save(oAuth2UserResponse.toUserEntity()));
+        userRepository.findByOauth2UserId(oAuth2UserId)
+            .orElseGet(() -> userRepository.save(oAuth2UserResponse.toUserEntity()));
 
         log.info("oAuth2UserId: {}", oAuth2UserResponse.getOAuth2UserId());
         return new CustomOAuth2User(oAuth2UserId, oAuth2UserResponse.getAttributes());
