--- service/src/main/java/uk/jinhy/server/service/vaccination/presentation/GlobalExceptionHandler.java
@@ -1,29 +1,0 @@
-package uk.jinhy.server.service.vaccination.presentation;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(PetNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handlePetNotFoundException(PetNotFoundException ex) {
-        ErrorResponse errorResponse = new ErrorResponse(
-            HttpStatus.NOT_FOUND.value(),
-            ex.getMessage()
-        );
-        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
-    }
-
-    // 다른 사용자 정의 예외에 대한 핸들러를 여기에 추가
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
-        ErrorResponse errorResponse = new ErrorResponse(
-            HttpStatus.INTERNAL_SERVER_ERROR.value(),
-            "예상치 못한 오류가 발생했습니다: " + ex.getMessage()
-        );
-        return new ResponseEntity<>(errorResponse, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
